isita_app/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
bleach==1.5.0
click==6.7
Flask==0.12.2
h5py==2.8.0
html5lib==0.9999999
intel-openmp==2018.0.0
itsdangerous==0.24
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
olefile==0.44
Pillow==5.3.0
pip
setuptools==36.5.0
torch
torchvision
tornado==4.5.2
urllib3==1.25.3
Werkzeug==0.12.2
wheel==0.30.0
numpy
